pyproject.toml
[project]
name = "omni-visual"
version = "0.1.0"
description = "Multimodal AI agent for blind accessibility - virtual eyes using Google Maps"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.35.0",
    "fastapi>=0.116.1",
    "google-genai>=1.36.0",
    "google-adk>=1.14.1",
    "httpx>=0.27.0",
    "mcp>=1.0.0",
    "tenacity>=8.2.0",
    "Pillow>=10.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/omni_visual"]
